Upgrade dependencies & refactor codes
