+++ .github/workflows/test_pull_request.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        php: ['8.1', '8.2', '8.3', '8.4', '8.5']
+        php: ['8.3', '8.4', '8.5']
 
     steps:
       - name: Set up PHP
@@ -24,7 +24,7 @@ jobs:
           node-version: '24.x'
 
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
         with:
           fetch-depth: 0
 
@@ -40,14 +40,6 @@ jobs:
 
       - name: PHP Lint
         run: ./vendor/bin/parallel-lint src tests examples
-
-      #- name: Neon Lint
-      #  run: ./vendor/nette/neon/bin/neon-lint conf
-
-      #- name: PHP MD
-      #  run: |
-      #    ./vendor/bin/phpmd --version
-      #    ./vendor/bin/phpmd ./src/ ./examples/ ./tests/ text phpmd.xml
 
       - name: PHP Code Sniffer
         run: |
@@ -71,7 +63,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        php: ['8.4']
+        php: ['8.5']
 
     steps:
       - name: Set up PHP
@@ -87,7 +79,7 @@ jobs:
           node-version: '24.x'
 
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
         with:
           fetch-depth: 0
 
@@ -95,7 +87,7 @@ jobs:
         run: composer install --no-interaction
 
       - name: Update PHPUnit for Code Coverage
-        run: composer require phpunit/phpunit:^10.5 sebastian/version:* --with-all-dependencies
+        run: composer require phpunit/phpunit:^12.5 sebastian/version:* --with-all-dependencies
 
       - name: PHP Lint
         run: ./vendor/bin/parallel-lint src tests examples
