+++ PHP_VERSIONS
@@ -1,5 +1,3 @@
-8.5
-8.4
 8.3
-8.2
-8.1
+8.4
+8.5
