+++ bin/TestAndLint.sh
@@ -33,9 +33,6 @@ test_and_lint() {
     echo "-----------------------------------------------------------"
     echo "[PHP $1][parallel-lint]"
     ./vendor/bin/parallel-lint src tests examples
-    #echo "-----------------------------------------------------------"
-    #echo "[PHP $1][neon-lint]"
-    #./vendor/nette/neon/bin/neon-lint conf
     echo "-----------------------------------------------------------"
     echo "[PHP $1][phpcs]"
     ./vendor/bin/phpcs --ignore=vendor \
@@ -43,11 +40,6 @@ test_and_lint() {
                         -p \
                         -s \
                         .
-    #echo "-----------------------------------------------------------"
-    #echo "[PHP $1][phpmd]"
-    #./vendor/bin/phpmd \
-    #                   ./src/ ./examples/ ./tests/ text \
-    #                   phpmd.xml
     echo "-----------------------------------------------------------"
     echo "[PHP $1][phpstan]"
     ./vendor/bin/phpstan analyze -c phpstan.neon
