+++ composer.json
@@ -17,12 +17,12 @@
     "minimum-stability": "stable",
     "require-dev": {
         "squizlabs/php_codesniffer": "^4.0",
-        "phpunit/phpunit": "^10.5",
+        "phpunit/phpunit": "^12.5",
         "phpstan/phpstan": "^2.1",
         "php-parallel-lint/php-parallel-lint": "^1.4",
-        "monolog/monolog": "^3.9"
+        "monolog/monolog": "^3.10"
     },
     "require": {
-        "php": ">=8.1"
+        "php": ">=8.3"
     }
 }
