+++ examples/UseInPhpUnitTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Macocci7\PhpCombination;
+namespace Macocci7\PhpCombination\Examples;
 
 require_once __DIR__ . '/../vendor/autoload.php';
 require_once __DIR__ . '/UseInPhpUnit.class.php';
@@ -30,11 +30,6 @@ final class UseInPhpUnitTest extends TestCase
         return $data;
     }
 
-    /**
-     * PHPDoc for PHPUnit 9.x
-     * @dataProvider provide_order_can_order_correctly
-     */
-    // Attribute for PHPUnit 10.x or later
     #[DataProvider('provide_order_can_order_correctly')]
     public function test_order_can_order_correctly(
         int $productId,
