--- phpmd.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<ruleset name="Custom_PHPMD">
-  <description>Custom ruleset PHPMD</description>
-  <!-- 実行するルールを追加する -->
-  <!-- ルール一覧 https://phpmd.org/rules/index.html -->
-  <rule ref="rulesets/cleancode.xml">
-    <!-- 除外するルール -->
-    <exclude name="MissingImport" />
-    <exclude name="StaticAccess" />
-    <exclude name="ElseExpression" />
-    <exclude name="BooleanArgumentFlag" />
-  </rule>
-  <rule ref="rulesets/codesize.xml">
-    <exclude name="TooManyPublicMethods" />
-  </rule>
-  <rule ref="rulesets/controversial.xml">
-    <exclude name="CamelCaseMethodName" />
-  </rule>
-  <rule ref="rulesets/design.xml" />
-  <rule ref="rulesets/unusedcode.xml" />
-</ruleset>
